Consolidate cri utility functions
